+++ apps/starter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # stoker
 
+## 0.4.212
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/web-app@0.5.177
+
 ## 0.4.211
 
 ### Patch Changes